requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Flask-Migrate
python-dotenv
email-validator
Flask-Mail
itsdangerous
python-dateutil>=2.8.2